Version 0.9.0

+++ source/Samples/GuiCommon/Src/RenderTargetTreeItemProperty.cpp
@@ -1,6 +1,6 @@
 #include "RenderTargetTreeItemProperty.hpp"
 
-#include <RenderTarget.hpp>
+#include <Render/RenderTarget.hpp>
 
 #include "AdditionalProperties.hpp"
 #include <wx/propgrid/advprops.h>
@@ -40,7 +40,7 @@ namespace GuiCommon
 				_( "Texture" )
 			};
 
-			p_grid->Append( new wxPropertyCategory( PROPERTY_CATEGORY_RENDER_TARGET + TARGETS[l_target->GetTargetType()] ) );
+			p_grid->Append( new wxPropertyCategory( PROPERTY_CATEGORY_RENDER_TARGET + TARGETS[size_t( l_target->GetTargetType() )] ) );
 		}
 	}
